Delete Engine class and move logic to FlowInstance model
- Engine class seemed redundant and logically it makes sense for a
  FlowInstance to be able to execute and update its own state.
- All the logic that was in the Engine class has been moved to the
  FlowInstance model along with the addition of some helper methods and
  properties.
- Added celery task to execute all the valid automatic transitions until
  no more can be executed. This task is then wrapped in another method
  which should be passed to the FlowInstance upon instantiation.
- Add simple tests to test basic flow logic and execution (no
  conditional logic added yet).
