+++ backend/app/models/flow_instance.py
@@ -1,8 +1,10 @@
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Callable, Optional
+from uuid import UUID
 
-from django.db import models
+from django.core.exceptions import ValidationError
+from django.db import models, transaction
 
-from app.models import Environment, FlowSchemaVersion, Person, StepSchema, TransitionSchema, UUIDModel
+from app.models import Environment, FlowSchema, FlowSchemaVersion, Person, StepSchema, TransitionSchema, UUIDModel
 
 if TYPE_CHECKING:
     from django.db.models.manager import RelatedManager
@@ -31,12 +33,232 @@ class FlowInstance(UUIDModel):
     steps: "RelatedManager[StepInstance]"
     transitions: "RelatedManager[TransitionInstance]"
 
+    class Meta:
+        constraints = [
+            # Ensure a person only has one active flow instance for a given schema version
+            models.UniqueConstraint(
+                fields=["person", "schema_version"],
+                name="unique_person_schema_version",
+            )
+        ]
+
+    # This is set via the factory constructor "start"
+    # It should be a callable that takes a FlowInstance ID and executes the automatic transitions for that flow
+    auto_transition_executor: Optional[Callable[[UUID], None]] = None
+
+    def call_auto_transition_executor(self):
+        """
+        Calls the automatic transitions task executor if it is set.
+
+        You need to pass in a shared celery task that will execute the automatic transitions for this to work as intended.
+        """
+        if self.state != self.FlowState.ACTIVE:
+            return
+        if self.auto_transition_executor and callable(self.auto_transition_executor):
+            self.auto_transition_executor(self.id)
+
+    def execute_automatic_transitions(self):
+        """
+        Executes all available automatic transitions for this flow instance.
+        """
+        if self.state != self.FlowState.ACTIVE:
+            return
+        with transaction.atomic():
+            for active_step in self.active_steps.all():
+                transition_count = 0
+                for transition in active_step.outgoing_transitions.filter(
+                    transition_schema__type=TransitionSchema.TransitionType.AUTOMATIC
+                ):
+                    if transition.can_transition:
+                        self.execute_manual_transition(transition)
+                        transition_count += 1
+                    else:
+                        continue
+                if transition_count:
+                    # Only want to set the active step to completed if at least one transition was executed
+                    active_step.state = StepInstance.StepState.COMPLETED
+                    active_step.save()
+        # Check if there are no more steps to execute
+        self._check_completed()
+
+    def execute_manual_transition(self, transition: "TransitionInstance | str", mark_as_completed: bool = False):
+        """
+        Executes the specified transition, moving the flow from one step to another.
+
+        Parameters:
+        - transition: The TransitionInstance to execute, or a TransitionSchema.identifier (str).
+        """
+        transition_instance: TransitionInstance | None = None
+        active_steps = self.active_steps.all()
+
+        if isinstance(transition, TransitionInstance):
+            transition_instance = transition
+        elif isinstance(transition, str):
+            # Find the TransitionInstance based on the active steps and the identifier
+            for active_step in active_steps:
+                transition_instance = TransitionInstance.objects.filter(
+                    step_instance_from=active_step,
+                    transition_schema__identifier=transition,
+                    flow_instance=self,
+                ).first()
+
+                if transition_instance:
+                    break  # Break the loop once the correct transition instance is found
+
+        if not transition_instance:
+            raise ValueError("Invalid transition argument or no active step matches the transition identifier.")
+
+        from_step = transition_instance.step_instance_from
+        to_step = transition_instance.step_instance_to
+
+        # Validate transition can be made
+        if from_step.state != StepInstance.StepState.ACTIVE:
+            raise ValidationError("Invalid transition state. The step you are transitioning from is not active.")
+        if to_step.state != StepInstance.StepState.INACTIVE:
+            raise ValidationError("Invalid transition state. The step you are transitioning to is not inactive.")
+
+        with transaction.atomic():
+            to_step.state = StepInstance.StepState.ACTIVE
+            to_step.save()
+
+            if mark_as_completed:
+                from_step.state = StepInstance.StepState.COMPLETED
+                from_step.save()
+
+        self._check_completed()
+
     @property
     def active_steps(self):
         """
         Returns a QuerySet of all StepInstance objects related to this FlowInstance that have the state 'ACTIVE'.
         """
         return self.steps.filter(state=StepInstance.StepState.ACTIVE)
+
+    @property
+    def has_automatic_transitions(self):
+        """
+        Returns True if there are any automatic transitions available for the current active steps, False otherwise.
+        """
+        return (
+            self.active_steps.filter(
+                outgoing_transitions__transition_schema__type=TransitionSchema.TransitionType.AUTOMATIC
+            ).count()
+            > 0
+        )
+
+    @property
+    def is_completed(self):
+        """
+        Returns True if the flow is completed, False otherwise.
+        """
+        if self.state == self.FlowState.COMPLETED:
+            return True
+        return self._check_completed()
+
+    def _check_completed(self) -> bool:
+        """
+        Checks if the flow is completed and updates the state if necessary.
+        """
+        # Assume all steps are completed.
+        all_steps_completed = True
+
+        for step in self.active_steps.all():
+            # If any step has outgoing transitions, mark as not all steps completed
+            if step.outgoing_transitions.exists():
+                all_steps_completed = False
+                break  # No need to check further if any step is not completed
+            else:
+                # Mark the individual step as completed
+                step.state = StepInstance.StepState.COMPLETED
+                step.save()
+
+        # Only if all steps are completed, mark the flow instance as completed
+        if all_steps_completed:
+            self.state = FlowInstance.FlowState.COMPLETED
+            self.save()
+
+        return all_steps_completed
+
+    @classmethod
+    def resume(
+        cls,
+        id: UUID,
+        auto_transition_executor: Optional[Callable[[UUID], None]] = None,
+    ) -> "FlowInstance":
+        """
+        Factory method to resume a flow given a FlowInstance ID.
+
+        Parameters:
+        - id: The UUID of the FlowInstance to resume.
+        """
+        flow_instance = cls.objects.get(id=id)
+        flow_instance.auto_transition_executor = auto_transition_executor
+        return flow_instance
+
+    @classmethod
+    def start(
+        cls,
+        person: Person,
+        flow_schema: FlowSchema,
+        auto_transition_executor: Optional[Callable[[UUID], None]] = None,
+    ) -> "FlowInstance":
+        """
+        Factory method to initialize and start a flow given a Person and FlowSchema.
+
+        Automatically sets up the flow instance, its steps, transitions, and executes any initial automatic transitions.
+
+        Parameters:
+        - person: The Person object for whom the flow is being started.
+        - flow_schema: The FlowSchema object defining the flow structure.
+        """
+
+        with transaction.atomic():
+            # Check for existing FlowInstance for this Person and FlowSchema
+            existing_instance = FlowInstance.objects.filter(
+                person=person,
+                schema_version__schema=flow_schema,
+            ).exists()
+            if existing_instance:
+                raise ValueError("A FlowInstance already exists for this Person and FlowSchema.")
+
+            latest_version = flow_schema.latest_version()
+            if not latest_version:
+                raise ValueError("No versions available for the provided FlowSchema.")
+
+            flow_instance = cls.objects.create(
+                person=person,
+                schema_version=latest_version,
+                environment=person.environment,
+            )
+
+            # Create StepInstances for each StepSchema in the FlowSchemaVersion
+            step_instances = {}
+            for step_schema in latest_version.steps.all():
+                step_instance = StepInstance.objects.create(
+                    step_schema=step_schema,
+                    flow_instance=flow_instance,
+                    state=StepInstance.StepState.INACTIVE,
+                )
+                step_instances[step_schema.id] = step_instance
+
+            # Create TransitionInstances for each TransitionSchema in the FlowSchemaVersion
+            for transition_schema in latest_version.transitions.all():
+                TransitionInstance.objects.create(
+                    transition_schema=transition_schema,
+                    flow_instance=flow_instance,
+                    step_instance_from=step_instances[transition_schema.from_step.id],
+                    step_instance_to=step_instances[transition_schema.to_step.id],
+                )
+
+            # Set the auto_transition_executor
+            # Don't want to call potentially long running tasks in a constructor so we just set it here
+            flow_instance.auto_transition_executor = auto_transition_executor
+
+            # Get the steps with no incoming transitions and set them as active (first steps in the flow)
+            first_steps = flow_instance.steps.filter(incoming_transitions=None).all()
+            first_steps.update(state=StepInstance.StepState.ACTIVE)
+
+            return flow_instance
 
 
 class StepInstance(UUIDModel):
@@ -76,3 +298,11 @@ class TransitionInstance(UUIDModel):
     flow_instance = models.ForeignKey(FlowInstance, on_delete=models.CASCADE, related_name="transitions")
     step_instance_from = models.ForeignKey(StepInstance, related_name="outgoing_transitions", on_delete=models.CASCADE)
     step_instance_to = models.ForeignKey(StepInstance, related_name="incoming_transitions", on_delete=models.CASCADE)
+
+    @property
+    def can_transition(self) -> bool:
+        """
+        Returns True if the transition can be executed, False otherwise.
+        """
+        # TODO: Evaluate conditional logic
+        return True
